[#113] Move floating label hooks to common utils
